backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pytz
astral
websockets
aiofiles
aiohttp
feedparser
paho-mqtt
jinja2
python-dotenv
psutil
